cascac

+++ ReceptsAPI/Controllers/CommentController.cs
@@ -106,11 +106,7 @@ namespace ReceptsAPI.Controllers
             {
                 return BadRequest();
             }
-            if (userId != userId && userId != request.Id) // схуяли нет юзер айди их реквеста, по моему написана хуйня
-            {
-                return BadRequest();
-
-            }
+            
             bool checkbool = _repository.Delete(userId);
             if (checkbool == false)
             {
@@ -134,8 +130,12 @@ namespace ReceptsAPI.Controllers
             {
                 return BadRequest();
             }
-            
 
+            if (userId != userId && userId != request.UserId) 
+            {
+                return BadRequest();
+
+            }
             bool checkbool = _repository.DeleteUser(userId);
             if (checkbool == false)
             {
